backend/requirements.txt
django==5.1
djangorestframework==3.15.2
django-cors-headers==4.4.0
gunicorn==22.0.0
